#290 🎉 feat: add RoutesIndexPage,validateQuerySearch and add helpers for SettingsPage
