+++ packages/ui/app/src/pages/RoutesPage/RoutesIndexPage.tsx
@@ -1,0 +1,8 @@
+import { EmptyState } from '@/components';
+
+export const RoutesIndexPage = () => (
+  <EmptyState
+    description='Pick a request from the list to inspect its routes'
+    title='No request selected'
+  />
+);
